refactor: ProductDetailActivity 리팩터링
