+++ app/src/main/res/layout/activity_product_detail.xml
@@ -7,7 +7,7 @@
 
         <variable
             name="presenter"
-            type="woowacourse.shopping.presentation.ui.productDetail.presenter.ProductDetailPresenter" />
+            type="woowacourse.shopping.presentation.ui.productDetail.ProductDetailPresenter" />
 
     </data>
